requirements.txt
httpx
rich
python-dotenv
jinja2